requirements.txt
attrs==23.1.0
autobahn==19.11.2
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.42.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiteconnect==4.2.0
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
nsepy==0.8
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
threadpoolctl==3.2.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
zope.interface==6.0
